This is an example of a C programme I wrote for a recent assignment. I achieved a 81.2 percent mark in this assignment.
This is an example of a C programme I wrote for a recent assignment. I was tasked with creating a programme which would take the results of a horticultural plant growing competition from a txt (text) file and store them within a dynamically allocated binary tree.

For part 1 of the assignment, had to print the results of the competition in order of longest total length of plant to shortest.