This is the second part of a C programme I wrote for a recent assignment. I achieved a 81.2 percent mark in this assignment.
This is an example of a C programme I wrote for a recent assignment. I was tasked with creating a programme which would take the results of a horticultural plant growing competition from a txt (text) file and store them within a dynamically allocated binary tree.

For part 2 of the assignment, we had to implement the option of also being able to print out the details of each competitor. This tested how well we could use pointers to avoid the use of global variables.